+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The minimum supported Python version is now 3.8.
 - Added support for PyTorch Lightning 1.7.x
 - The Beaker Executor will no-longer live-stream logs from Beaker jobs, but logs will be viewable on Beaker and more readable.
+- Only the Beaker executor requires a clean working directory
 
 ### Fixed